requirements.txt
Django>=2.1.3,<2.2.0
django-cors-headers==3.5.0
django-rest-swagger==2.2.0
django-filter==2.4.0
djangorestframework>=3.9.0,<3.10.0
mysqlclient==2.0.3
flake8>=3.6.0,<3.7.0